+++ include/IDrawable.hpp
@@ -7,7 +7,7 @@ public:
 	IDrawable* m_pNextDrawable{nullptr};
 
 	IDrawable& operator=(const IDrawable&) = delete;
-	virtual bool Draw(float posX, float posY, float deltaTime) = 0;
+	virtual bool Draw(int posX, int posY, float deltaTime) = 0;
 	virtual ~IDrawable() = default;
 };
